Add scanCore, a CLI virus scanner in PHP.
The next iteration of PHP-AV is here! This version is a high-performance, single-threaded CLI (command-line) version for use in the HR-AV app. However, it works great as a standalone tool as well! 

To use it, just call a PHP binary with the path to this script and it's arguments. Presto! You're off scanning files.